- fb_accel_cs_hd1: fix compile warning

+++ src/driver/fb_accel_cs_hd1.cpp
@@ -288,9 +288,9 @@ void CFbAccelCSHD1::blit2FB(void *fbbuff, uint32_t width, uint32_t height, uint3
 	yc = (height > yRes) ? yRes : height;
 
 	//NI
-	if (unscaled_w != 0 && (int)unscaled_w < xc)
+	if (unscaled_w != 0 && unscaled_w < xc)
 		xc = unscaled_w;
-	if (unscaled_h != 0 && (int)unscaled_h < yc)
+	if (unscaled_h != 0 && unscaled_h < yc)
 		yc = unscaled_h;
 
 	u32 cmd;
